Added eclipse test case and designed more clean approach for base class
