+++ src/fixtures/ide_fixture.py
@@ -1,7 +1,8 @@
 import json
 import pytest
 import os
-from src.lib.codereadystudio import CodeReadyStudio
+from src.lib.application import CodeReadyStudio
+from src.lib.application import Eclipse
 
 
 CONF_DIR = os.path.dirname(os.path.dirname(
@@ -21,11 +22,25 @@ def config():
 @pytest.fixture(scope='function')
 def setup_codereadystudio(config):
     """
-    Fixture to setup codereadystudio class
+    Fixture to setup codereadystudio application
     """
     codereadystudio = CodeReadyStudio()
     path = config['ide_paths']['rh_code_ready_studio']
-    codereadystudio_app = codereadystudio.open_application(path)
+    codereadystudio.open_application(path)
     codereadystudio.set_default_timeout(timeout=config['timeout_in_seconds'])
     yield codereadystudio
-    codereadystudio.close_application(codereadystudio_app)
+    # codereadystudio.close_application(codereadystudio_app)
+    codereadystudio.close_ide()
+
+
+@pytest.fixture(scope='function')
+def setup_eclipse(config):
+    """
+    Fixture to setup eclipse application
+    """
+    eclipse = Eclipse()
+    path = config['ide_paths']['eclipse']
+    eclipse.open_application(path)
+    eclipse.set_default_timeout(timeout=config['timeout_in_seconds'])
+    yield eclipse
+    eclipse.close_ide()
