+++ src/images/codeready_eclipse_common/generating_report.png
