+++ src/images/codeready_eclipse_common/ok_button.png
