+++ src/images/codeready_eclipse_common/project_highlighted.png
