+++ src/images/codeready_eclipse_common/run_complete.png
