+++ src/lib/application.py
@@ -1,9 +1,9 @@
 from RPA.Desktop import Desktop
 import os
-import src.lib.codereadystudio as codereadystudio
+import logging
 
 
-class BaseDesktop(Desktop):
+class Application(Desktop):
     """
         Library for customizing the basic Desktop class
         functionalities adapted for generic usage
@@ -22,8 +22,8 @@ class BaseDesktop(Desktop):
         Returns:
             (str): Image locator string
         """
-        if isinstance(self, codereadystudio.CodeReadyStudio):
-            return f'image:{self.IMG_DIR}/codereadystudio/{locator}'
+        if isinstance(self, CodeReadyStudio) or isinstance(self, Eclipse):
+            return f'image:{self.IMG_DIR}/codeready_eclipse_common/{locator}'
 
     def text_locator(self, locator):
         """
@@ -168,3 +168,125 @@ class BaseDesktop(Desktop):
                               timeout=timeout, interval=interval)
         element_position = self.find_element(locator=formed_locator)
         return element_position
+
+    def close_ide(self):
+        """
+        Closes the IDE
+        """
+        self.click_element(locator_type='text', locator='File')
+        self.click_element(locator_type='text', locator='Exit')
+
+    def is_open_mta_perspective(self):
+        """
+        Checks if MTA perspective is already opened in IDE
+
+        Returns:
+            (bool): True or False
+        """
+        try:
+            self.wait_find_element(locator_type='image',
+                                   locator='mta_perspective.png')
+            return True
+        except Exception as exc:
+            logging.debug('An error occured while finding \
+                MTA perspective tab ! {}'.format(str(exc)))
+            if 'No matches found' in str(exc):
+                return False
+            else:
+                raise Exception(exc)
+
+    def open_mta_perspective(self):
+        """
+        Opens MTA perspective in IDE
+        """
+        if self.is_open_mta_perspective():
+            logging.info('MTA perspective is already opened !')
+            return
+        else:
+            """
+            Click on Window -> Perspective -> Open Perspective
+            -> Other -> MTA -> Click on Open button
+            """
+            self.click_element(locator_type='text', locator='Window')
+            self.click_element(locator_type='text', locator='Perspective')
+            self.click_element(locator_type='image',
+                               locator='open_perspective.png')
+            try:
+                self.click_element(locator_type='text', locator='Other...')
+            except Exception as exc:
+                # Re-attempt to click 'Other' using image locator
+                if 'No matches found' in str(exc):
+                    self.click_element(locator_type='image',
+                                       locator='other_alt.png')
+                else:
+                    raise Exception(exc)
+            self.click_element(locator_type='image',
+                               locator='open_perspective_header.png')
+            self.type_text(text='mta')
+            open_perspective_button = self.image_locator('open_button.png')
+            self.move_mouse(open_perspective_button)
+            self.click(action='double_click')
+
+    def run_simple_analysis(self, project, packages=[]):
+        """
+        Runs analysis by adding the project and/or packages passed as argument
+
+        Args:
+            project (str): Full name of project to be analysed
+            packages (list): List of packages to be added to analysis
+
+        Returns:
+            None
+        """
+        """
+        Click on Run MTA Configuration icon -> Add Project button
+        -> Type Project name -> Click on OK button
+        Confirm analysis has started by finding generating report
+        """
+        self.click_element(locator_type='image', locator='mta_run.png')
+        self.click_element(locator_type='image',
+                           locator='run_conf_header.png')
+        # Get the multiple Add Project Buttons and select first one
+        add_button_locator = self.image_locator('add_button.png')
+        add_buttons = self.find_elements(add_button_locator)
+        self.click(locator=add_buttons[0])
+        self.click_element(locator_type='image',
+                           locator='projects_header.png')
+        self.type_text(text=project, enter=True)
+        # Disable generate report from options
+        self.click_element(locator_type='text', locator='Options')
+        self.click_element(locator_type='image',
+                           locator='generate_report_checkbox.png')
+        self.click_element(locator_type='image',
+                           locator='run_config_button.png')
+        self.wait_find_element(locator_type='image',
+                               locator='generating_report.png')
+        self.wait_find_element(locator_type='image',
+                               locator='run_complete.png',
+                               timeout=120.0, interval=5.0)
+
+    def is_analysis_complete(self):
+        """
+        Checks if run analysis has been completed
+
+        Returns:
+            bool: True if analysis was completed
+        """
+        self.wait_find_element(locator_type='image',
+                               locator='run_complete.png',
+                               timeout=120.0, interval=5.0)
+        return True
+
+
+class CodeReadyStudio(Application):
+    """
+    CodeReadyStudio class for managing RH Code Ready Studio application
+    """
+    pass
+
+
+class Eclipse(Application):
+    """
+    Eclipse class for managing Eclipse application
+    """
+    pass
